@@ -1,20 +1,26 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
     <AssemblyName>PHP.WebServiceConcept.AccountWebService</AssemblyName>
     <RootNamespace>PHP.WebServiceConcept.AccountWebService</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Model\" />
-    <Folder Include="wwwroot\" />
+    <Compile Remove="Model\**" />
+    <Compile Remove="wwwroot\**" />
+    <Content Remove="Model\**" />
+    <Content Remove="wwwroot\**" />
+    <EmbeddedResource Remove="Model\**" />
+    <EmbeddedResource Remove="wwwroot\**" />
+    <None Remove="Model\**" />
+    <None Remove="wwwroot\**" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.0" />
-    <PackageReference Include="SimpleInjector" Version="4.0.11" />
-    <PackageReference Include="SimpleInjector.Integration.AspNetCore.Mvc" Version="4.0.11" />
+    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.3" />
+    <PackageReference Include="SimpleInjector" Version="4.3.0" />
+    <PackageReference Include="SimpleInjector.Integration.AspNetCore.Mvc" Version="4.3.1" />
   </ItemGroup>
 
   <ItemGroup>
